refactor: split configeditor/fields_system.go by sub-screen
841 lines holding every System Setup and Access & Security sub-screen's field
definitions, split along the two inner menus the TUI already presents:

  fields_system.go            107  timezone lookup, buildSysFields, both menus
  fields_system_setup.go      110  registration, defaults, DOS emulation
  fields_system_network.go    174  server setup (listeners, hostnames)
  fields_system_logging.go    113  logging + its log-rolling type lookup
  fields_security.go          200  access levels, connection limits, IP lists
  fields_security_defense.go  189  bot defense, new-user voting

Pure move: no declaration bodies changed. Verified with a go/parser pairwise
diff of every top-level declaration, printed and compared — 18 old, 18 new,
0 missing, 0 mismatched, 0 extra. No test file touched.
