+++ internal/configeditor/fields_system.go
@@ -1,11 +1,8 @@
 package configeditor
 
-import (
-	"github.com/ViSiON-3/vision-3-bbs/internal/uitext"
-	"strconv"
-
-	"github.com/ViSiON-3/vision-3-bbs/internal/config"
-)
+// Field definitions for the System Setup and Access & Security inner menus.
+// The sysFields* builders themselves live in the fields_system_*/fields_security*
+// files named for the sub-screen they build.
 
 // buildTimezoneLookupItems returns a list of common IANA timezones.
 func buildTimezoneLookupItems() []LookupItem {
@@ -106,736 +103,5 @@ func securityMenuItems() []sysConfigMenuItem {
 		{Label: "Bot Defense", Build: func(m *Model) []fieldDef { return sysFieldsBotDefense(&m.configs.Server) }},
 		{Label: "IP Blocklist/Allowlist", Build: func(m *Model) []fieldDef { return sysFieldsIPLists(&m.configs.Server) }},
 		{Label: "New User Voting (NUV)", Build: func(m *Model) []fieldDef { return sysFieldsNUV(&m.configs.Server) }},
-	}
-}
-
-// sysFieldsBotDefense returns fields for the Bot Defense sub-screen
-// (Challenge Gate + connection-rate limiter).
-func sysFieldsBotDefense(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "Enable Gate", Help: "Require a key-press challenge before login", Type: ftYesNo, Col: 3, Row: 1, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.EnableChallengeGate) },
-			Set: func(val string) error { cfg.EnableChallengeGate = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Gate Art File", Help: "ANSI/ASCII file in the menu ansi dir", Type: ftString, Col: 3, Row: 2, Width: 20,
-			Get: func() string { return cfg.ChallengeGateFile },
-			Set: func(val string) error { cfg.ChallengeGateFile = val; return nil },
-		},
-		{
-			Label: "Challenge Key", Help: "ESC or a single character to press", Type: ftString, Col: 3, Row: 3, Width: 6,
-			Get: func() string { return cfg.ChallengeGateKey },
-			Set: func(val string) error { cfg.ChallengeGateKey = val; return nil },
-		},
-		{
-			Label: "Timeout Secs", Help: "Seconds allowed to complete the challenge", Type: ftInteger, Col: 3, Row: 4, Width: 4, Min: 1, Max: 999,
-			Get: func() string { return strconv.Itoa(cfg.ChallengeGateTimeoutSeconds) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.ChallengeGateTimeoutSeconds = n
-				return nil
-			},
-		},
-		{
-			Label: "Req Presses", Help: "Key presses required to pass", Type: ftInteger, Col: 3, Row: 5, Width: 3, Min: 1, Max: 99,
-			Get: func() string { return strconv.Itoa(cfg.ChallengeGateRequiredPresses) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.ChallengeGateRequiredPresses = n
-				return nil
-			},
-		},
-		{
-			Label: "Live Countdown", Help: "Animate the ## countdown (off = static)", Type: ftYesNo, Col: 3, Row: 6, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.ChallengeGateLiveCountdown) },
-			Set: func(val string) error { cfg.ChallengeGateLiveCountdown = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Rate Limit", Help: "Temp-ban IPs that reconnect too fast", Type: ftYesNo, Col: 3, Row: 7, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.EnableConnRateLimit) },
-			Set: func(val string) error { cfg.EnableConnRateLimit = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Rate Hits", Help: "Attempts within the window that trigger a ban", Type: ftInteger, Col: 3, Row: 8, Width: 4, Min: 0, Max: 9999,
-			Get: func() string { return strconv.Itoa(cfg.ConnRateLimitHits) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.ConnRateLimitHits = n
-				return nil
-			},
-		},
-		{
-			Label: "Rate Window", Help: "Sliding window in seconds", Type: ftInteger, Col: 3, Row: 9, Width: 4, Min: 1, Max: 9999,
-			Get: func() string { return strconv.Itoa(cfg.ConnRateLimitWindowSeconds) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.ConnRateLimitWindowSeconds = n
-				return nil
-			},
-		},
-		{
-			Label: "Ban Minutes", Help: "Temp-ban duration in minutes", Type: ftInteger, Col: 3, Row: 10, Width: 5, Min: 1, Max: 99999,
-			Get: func() string { return strconv.Itoa(cfg.ConnRateLimitBanMinutes) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.ConnRateLimitBanMinutes = n
-				return nil
-			},
-		},
-	}
-}
-
-// sysFieldsRegistration returns fields for BBS Registration sub-screen.
-func sysFieldsRegistration(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "Board Name", Help: "Your BBS name shown to users", Type: ftString, Col: 3, Row: 1, Width: 40,
-			Get: func() string { return cfg.BoardName },
-			Set: func(val string) error { cfg.BoardName = val; return nil },
-		},
-		{
-			Label: "SysOp Name", Help: "System operator name", Type: ftString, Col: 3, Row: 2, Width: 30,
-			Get: func() string { return cfg.SysOpName },
-			Set: func(val string) error { cfg.SysOpName = val; return nil },
-		},
-		{
-			Label: "BBS Location", Help: "Geographic location (e.g. Auckland, NZ)", Type: ftString, Col: 3, Row: 3, Width: 40,
-			Get: func() string { return cfg.BBSLocation },
-			Set: func(val string) error { cfg.BBSLocation = val; return nil },
-		},
-		{
-			Label: "Timezone", Help: "IANA timezone", Type: ftLookup, Col: 3, Row: 4, Width: 30,
-			Get: func() string { return cfg.Timezone },
-			Set: func(val string) error { cfg.Timezone = val; return nil },
-			LookupItems: func() []LookupItem {
-				items := buildTimezoneLookupItems()
-
-				// If current timezone is not empty and not in the curated list, append it
-				if cfg.Timezone != "" {
-					found := false
-					for _, item := range items {
-						if item.Value == cfg.Timezone || item.Display == cfg.Timezone {
-							found = true
-							break
-						}
-					}
-					if !found {
-						items = append(items, LookupItem{
-							Value:   cfg.Timezone,
-							Display: cfg.Timezone + " (current)",
-						})
-					}
-				}
-
-				return items
-			},
-		},
-		{
-			Label: "QWK ID", Help: "Stable QWK packet ID (max 8, A-Z/0-9); blank = derive from Board Name",
-			Type: ftString, Col: 3, Row: 5, Width: 8,
-			Get: func() string { return cfg.QWKID },
-			Set: func(val string) error { cfg.QWKID = config.NormalizeQWKID(val); return nil },
-		},
-	}
-}
-
-// sysFieldsNetwork returns fields for Server Setup sub-screen.
-func (m *Model) sysFieldsNetwork(cfg *config.ServerConfig) []fieldDef {
-	v3 := &m.configs.V3Net
-	hub := &m.configs.V3Net.Hub
-	binkd := &m.configs.FTN.Binkd
-
-	return []fieldDef{
-		{
-			Label: "SSH Enabled", Help: "Enable SSH server", Type: ftYesNo, Col: 3, Row: 1, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.SSHEnabled) },
-			Set: func(val string) error { cfg.SSHEnabled = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "SSH Host", Help: "Listen address (blank=all interfaces)", Type: ftString, Col: 3, Row: 2, Width: 20,
-			Get: func() string { return cfg.SSHHost },
-			Set: func(val string) error { cfg.SSHHost = val; return nil },
-		},
-		{
-			Label: "SSH Port", Help: "SSH listen port (default: 8022)", Type: ftInteger, Col: 3, Row: 3, Width: 5, Min: 1, Max: 65535,
-			Get: func() string { return strconv.Itoa(cfg.SSHPort) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.SSHPort = n
-				return nil
-			},
-		},
-		{
-			Label: "Legacy SSH", Help: "Allow legacy algorithms for older clients", Type: ftYesNo, Col: 3, Row: 4, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.LegacySSHAlgorithms) },
-			Set: func(val string) error { cfg.LegacySSHAlgorithms = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Telnet Enabled", Help: "Enable Telnet server", Type: ftYesNo, Col: 3, Row: 6, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.TelnetEnabled) },
-			Set: func(val string) error { cfg.TelnetEnabled = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Telnet Host", Help: "Listen address (blank=all interfaces)", Type: ftString, Col: 3, Row: 7, Width: 20,
-			Get: func() string { return cfg.TelnetHost },
-			Set: func(val string) error { cfg.TelnetHost = val; return nil },
-		},
-		{
-			Label: "Telnet Port", Help: "Telnet listen port (default: 8023)", Type: ftInteger, Col: 3, Row: 8, Width: 5, Min: 1, Max: 65535,
-			Get: func() string { return strconv.Itoa(cfg.TelnetPort) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.TelnetPort = n
-				return nil
-			},
-		},
-		{
-			Label: "V3Net", Help: "Enable V3Net networking", Type: ftYesNo, Col: 3, Row: 10, Width: 1,
-			Get: func() string { return uitext.BoolToYN(v3.Enabled) },
-			Set: func(val string) error { v3.Enabled = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Keystore Path", Help: "Path to Ed25519 keypair file", Type: ftString, Col: 3, Row: 11, Width: 40,
-			Get: func() string { return v3.KeystorePath },
-			Set: func(val string) error { v3.KeystorePath = val; return nil },
-		},
-		{
-			Label: "Dedup DB Path", Help: "Path to deduplication SQLite database", Type: ftString, Col: 3, Row: 12, Width: 40,
-			Get: func() string { return v3.DedupDBPath },
-			Set: func(val string) error { v3.DedupDBPath = val; return nil },
-		},
-		{
-			Label: "Registry URL", Help: "Central V3Net registry URL (optional)", Type: ftString, Col: 3, Row: 13, Width: 49,
-			Get: func() string { return v3.RegistryURL },
-			Set: func(val string) error { v3.RegistryURL = val; return nil },
-		},
-		{
-			Label: "V3Net Hub", Help: "Run a V3Net hub server on this node", Type: ftYesNo, Col: 3, Row: 15, Width: 1,
-			Get: func() string { return uitext.BoolToYN(hub.Enabled) },
-			Set: func(val string) error { hub.Enabled = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Hub Host", Help: "Hub listen address (blank=all interfaces)", Type: ftString, Col: 3, Row: 16, Width: 20,
-			Get: func() string { return hub.Host },
-			Set: func(val string) error { hub.Host = val; return nil },
-		},
-		{
-			Label: "Hub Port", Help: "Hub listen port (default: 8765)", Type: ftInteger, Col: 3, Row: 17, Width: 5, Min: 1, Max: 65535,
-			Get: func() string {
-				p := hub.Port
-				if p == 0 {
-					p = 8765
-				}
-				return strconv.Itoa(p)
-			},
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				hub.Port = n
-				return nil
-			},
-		},
-		{
-			Label: "Hub Data Dir", Help: "Hub data storage directory", Type: ftString, Col: 3, Row: 18, Width: 40,
-			Get: func() string { return hub.DataDir },
-			Set: func(val string) error { hub.DataDir = val; return nil },
-		},
-		{
-			Label: "Auto Approve", Help: "Automatically approve new leaf subscriptions", Type: ftYesNo, Col: 3, Row: 19, Width: 1,
-			Get: func() string { return uitext.BoolToYN(hub.AutoApprove) },
-			Set: func(val string) error { hub.AutoApprove = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Binkd Mailer", Help: "Run bundled binkd FTN mailer at startup", Type: ftYesNo, Col: 3, Row: 21, Width: 1,
-			Get: func() string { return uitext.BoolToYN(binkd.Enabled) },
-			Set: func(val string) error { binkd.Enabled = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Binkd Port", Help: "binkp listen port (default: 24554)", Type: ftInteger, Col: 3, Row: 22, Width: 5, Min: 1, Max: 65535,
-			Get: func() string { return strconv.Itoa(binkd.Port) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				binkd.Port = n
-				return nil
-			},
-		},
-		{
-			Label: "Binkd Binary", Help: "Path to binkd binary (default: bin/binkd)", Type: ftString, Col: 3, Row: 23, Width: 40,
-			Get: func() string { return binkd.BinaryPath },
-			Set: func(val string) error { binkd.BinaryPath = val; return nil },
-		},
-		{
-			Label: "Binkd Log Lvl", Help: "binkd loglevel 1-9 (default: 4)", Type: ftInteger, Col: 3, Row: 24, Width: 2, Min: 1, Max: 9,
-			Get: func() string { return strconv.Itoa(binkd.LogLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				binkd.LogLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "Export Secs", Help: "Outbound scan/pack interval in seconds (default: 300)", Type: ftInteger, Col: 3, Row: 25, Width: 6, Min: 30, Max: 86400,
-			Get: func() string { return strconv.Itoa(binkd.ExportSecs) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				binkd.ExportSecs = n
-				return nil
-			},
-		},
-	}
-}
-
-// sysFieldsLimits returns fields for Connection Limits sub-screen.
-func sysFieldsLimits(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "Max Nodes", Help: "Maximum simultaneous connections", Type: ftInteger, Col: 3, Row: 1, Width: 5, Min: 1, Max: 999,
-			Get: func() string { return strconv.Itoa(cfg.MaxNodes) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.MaxNodes = n
-				return nil
-			},
-		},
-		{
-			Label: "Max Per IP", Help: "Max connections from a single IP address", Type: ftInteger, Col: 3, Row: 2, Width: 5, Min: 1, Max: 999,
-			Get: func() string { return strconv.Itoa(cfg.MaxConnectionsPerIP) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.MaxConnectionsPerIP = n
-				return nil
-			},
-		},
-		{
-			Label: "Failed Logins", Help: "Failed attempts before lockout (0=disabled)", Type: ftInteger, Col: 3, Row: 3, Width: 5, Min: 0, Max: 100,
-			Get: func() string { return strconv.Itoa(cfg.MaxFailedLogins) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.MaxFailedLogins = n
-				return nil
-			},
-		},
-		{
-			Label: "Lockout Mins", Help: "Lockout duration after failed logins", Type: ftInteger, Col: 3, Row: 4, Width: 5, Min: 0, Max: 9999,
-			Get: func() string { return strconv.Itoa(cfg.LockoutMinutes) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.LockoutMinutes = n
-				return nil
-			},
-		},
-		{
-			Label: "Idle Timeout", Help: "Disconnect idle users after N minutes", Type: ftInteger, Col: 3, Row: 5, Width: 5, Min: 0, Max: 999,
-			Get: func() string { return strconv.Itoa(cfg.SessionIdleTimeoutMinutes) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.SessionIdleTimeoutMinutes = n
-				return nil
-			},
-		},
-		{
-			Label: "Xfer Timeout", Help: "Max minutes for active file transfers (0=unlimited)", Type: ftInteger, Col: 3, Row: 6, Width: 5, Min: 0, Max: 999,
-			Get: func() string { return strconv.Itoa(cfg.TransferTimeoutMinutes) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.TransferTimeoutMinutes = n
-				return nil
-			},
-		},
-	}
-}
-
-// sysFieldsLevels returns fields for Access Levels sub-screen.
-func sysFieldsLevels(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "SysOp Level", Help: "Security level for full SysOp access", Type: ftInteger, Col: 3, Row: 1, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.SysOpLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.SysOpLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "CoSysOp Level", Help: "Security level for CoSysOp access", Type: ftInteger, Col: 3, Row: 2, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.CoSysOpLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.CoSysOpLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "WFC Access", Help: "Allow remote WFC sysop console connections", Type: ftYesNo, Col: 3, Row: 3, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.WFCEnabled) },
-			Set: func(val string) error { cfg.WFCEnabled = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Invisible Lvl", Help: "Level at which user is hidden from who's online", Type: ftInteger, Col: 3, Row: 4, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.InvisibleLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.InvisibleLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "New User Level", Help: "Level assigned to new signups", Type: ftInteger, Col: 3, Row: 5, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.NewUserLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.NewUserLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "Regular Level", Help: "Level assigned when user is validated", Type: ftInteger, Col: 3, Row: 6, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.RegularUserLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.RegularUserLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "Logon Level", Help: "Minimum access level required to log in (0=disabled)", Type: ftInteger, Col: 3, Row: 7, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.LogonLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.LogonLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "Anonymous Lvl", Help: "Minimum level required to post anonymously (0=disabled)", Type: ftInteger, Col: 3, Row: 8, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.AnonymousLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.AnonymousLevel = n
-				return nil
-			},
-		},
-	}
-}
-
-// sysFieldsDefaults returns fields for Default Settings sub-screen.
-func sysFieldsDefaults(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "Allow New Users", Help: "Allow new user registration", Type: ftYesNo, Col: 3, Row: 1, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.AllowNewUsers) },
-			Set: func(val string) error { cfg.AllowNewUsers = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "File List Mode", Help: "File listing style", Type: ftLookup, Col: 3, Row: 2, Width: 15,
-			Get: func() string { return cfg.FileListingMode },
-			Set: func(val string) error { cfg.FileListingMode = val; return nil },
-			LookupItems: func() []LookupItem {
-				return []LookupItem{
-					{Value: "lightbar", Display: "lightbar - Interactive arrow-key navigation"},
-					{Value: "classic", Display: "classic - Traditional numbered list"},
-				}
-			},
-		},
-		{
-			Label: "Del User Days", Help: "Days to keep deleted user records (0=purge now, -1=forever)", Type: ftInteger, Col: 3, Row: 3, Width: 5, Min: -1, Max: 9999,
-			Get: func() string { return strconv.Itoa(cfg.DeletedUserRetentionDays) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.DeletedUserRetentionDays = n
-				return nil
-			},
-		},
-	}
-}
-
-// sysFieldsIPLists returns fields for IP Blocklist/Allowlist sub-screen.
-func sysFieldsIPLists(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "Blocklist Path", Help: "Path to IP blocklist file (one IP per line)", Type: ftString, Col: 3, Row: 1, Width: 45,
-			Get: func() string { return cfg.IPBlocklistPath },
-			Set: func(val string) error { cfg.IPBlocklistPath = val; return nil },
-		},
-		{
-			Label: "Allowlist Path", Help: "Path to IP allowlist file (one IP per line)", Type: ftString, Col: 3, Row: 2, Width: 45,
-			Get: func() string { return cfg.IPAllowlistPath },
-			Set: func(val string) error { cfg.IPAllowlistPath = val; return nil },
-		},
-	}
-}
-
-// sysFieldsNUV returns fields for the New User Voting sub-screen.
-func sysFieldsNUV(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "Use NUV", Help: "Enable New User Voting system", Type: ftYesNo, Col: 3, Row: 1, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.UseNUV) },
-			Set: func(val string) error { cfg.UseNUV = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Auto Add NUV", Help: "Automatically queue new registrations for voting", Type: ftYesNo, Col: 3, Row: 2, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.AutoAddNUV) },
-			Set: func(val string) error { cfg.AutoAddNUV = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "NUV Use Level", Help: "Minimum access level required to vote", Type: ftInteger, Col: 3, Row: 3, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.NUVUseLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.NUVUseLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "NUV Yes Votes", Help: "YES votes required to trigger yes threshold", Type: ftInteger, Col: 3, Row: 4, Width: 3, Min: 1, Max: 999,
-			Get: func() string { return strconv.Itoa(cfg.NUVYesVotes) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.NUVYesVotes = n
-				return nil
-			},
-		},
-		{
-			Label: "NUV No Votes", Help: "NO votes required to trigger no threshold", Type: ftInteger, Col: 3, Row: 5, Width: 3, Min: 1, Max: 999,
-			Get: func() string { return strconv.Itoa(cfg.NUVNoVotes) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.NUVNoVotes = n
-				return nil
-			},
-		},
-		{
-			Label: "NUV Validate", Help: "Auto-validate user when yes threshold reached", Type: ftYesNo, Col: 3, Row: 6, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.NUVValidate) },
-			Set: func(val string) error { cfg.NUVValidate = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "NUV Kill", Help: "Auto-delete user when no threshold reached", Type: ftYesNo, Col: 3, Row: 7, Width: 1,
-			Get: func() string { return uitext.BoolToYN(cfg.NUVKill) },
-			Set: func(val string) error { cfg.NUVKill = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "NUV Level", Help: "Access level assigned to auto-validated NUV users", Type: ftInteger, Col: 3, Row: 8, Width: 3, Min: 0, Max: 255,
-			Get: func() string { return strconv.Itoa(cfg.NUVLevel) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.NUVLevel = n
-				return nil
-			},
-		},
-		{
-			Label: "NUV Form", Help: "Infoform number (1-5) shown to voters; 0 = disabled", Type: ftInteger, Col: 3, Row: 9, Width: 1, Min: 0, Max: 5,
-			Get: func() string { return strconv.Itoa(cfg.NUVForm) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				cfg.NUVForm = n
-				return nil
-			},
-		},
-	}
-}
-
-// sysFieldsDOS returns fields for DOS Emulation sub-screen.
-func sysFieldsDOS(cfg *config.ServerConfig) []fieldDef {
-	return []fieldDef{
-		{
-			Label: "DOSemu Path", Help: "Path to dosemu2 binary (blank=auto-detect)", Type: ftString, Col: 3, Row: 1, Width: 45,
-			Get: func() string { return cfg.DosemuPath },
-			Set: func(val string) error { cfg.DosemuPath = val; return nil },
-		},
-	}
-}
-
-// logRollingTypes is the single source of truth for the Rolling Type lookup:
-// the persisted LoggingConfig.Type, the short label shown in the form, and the
-// picker description. logTypeToLabel / logLabelToType convert against it.
-var logRollingTypes = []struct {
-	typ     int
-	label   string
-	display string
-}{
-	{config.LogTypeNone, "None", "None - single file, no rotation"},
-	{config.LogTypeSize, "Size", "Size - rotate at Max Size KB, keep Max Files"},
-	{config.LogTypeDaily, "Daily", "Daily - new file per day, keep Max Files days"},
-}
-
-// logTypeToLabel maps a persisted type to its label, falling back to "None" for
-// an unmapped/hand-edited value (the rolling writer treats unknown types as
-// no-rotation, so "None" is the consistent display).
-func logTypeToLabel(typ int) string {
-	for _, t := range logRollingTypes {
-		if t.typ == typ {
-			return t.label
-		}
-	}
-	return "None"
-}
-
-func logLabelToType(label string) int {
-	for _, t := range logRollingTypes {
-		if t.label == label {
-			return t.typ
-		}
-	}
-	return config.LogTypeNone
-}
-
-func logRollingTypeItems() []LookupItem {
-	items := make([]LookupItem, len(logRollingTypes))
-	for i, t := range logRollingTypes {
-		items[i] = LookupItem{Value: t.label, Display: t.display}
-	}
-	return items
-}
-
-// sysFieldsLogging returns fields for the Logging sub-screen.
-func sysFieldsLogging(cfg *config.ServerConfig) []fieldDef {
-	lg := &cfg.Logging
-	return []fieldDef{
-		{
-			Label: "Log Directory", Help: "Directory for log files (absolute, or relative to BBS root)", Type: ftString, Col: 3, Row: 1, Width: 40,
-			Get: func() string { return lg.Dir },
-			Set: func(val string) error { lg.Dir = val; return nil },
-		},
-		{
-			Label: "Min Level", Help: "Minimum level written to the log", Type: ftLookup, Col: 3, Row: 2, Width: 8,
-			Get: func() string { return lg.Level },
-			Set: func(val string) error { lg.Level = val; return nil },
-			LookupItems: func() []LookupItem {
-				return []LookupItem{
-					{Value: "DEBUG", Display: "DEBUG - most verbose"},
-					{Value: "INFO", Display: "INFO - normal operation"},
-					{Value: "WARN", Display: "WARN - warnings and errors"},
-					{Value: "ERROR", Display: "ERROR - errors only"},
-				}
-			},
-		},
-		{
-			Label: "Rolling Type", Help: "How log files are rotated", Type: ftLookup, Col: 3, Row: 3, Width: 8,
-			Get:         func() string { return logTypeToLabel(lg.Type) },
-			Set:         func(val string) error { lg.Type = logLabelToType(val); return nil },
-			LookupItems: logRollingTypeItems,
-		},
-		{
-			Label: "Cache Writes", Help: "Buffer writes in an 8KB cache (flushed on errors/exit)", Type: ftYesNo, Col: 3, Row: 4, Width: 1,
-			Get: func() string { return uitext.BoolToYN(lg.Cache) },
-			Set: func(val string) error { lg.Cache = uitext.YNToBool(val); return nil },
-		},
-		{
-			Label: "Max Files", Help: "Retained backups (Size type) or days (Daily type)", Type: ftInteger, Col: 3, Row: 5, Width: 5, Min: 1, Max: 9999,
-			Get: func() string { return strconv.Itoa(lg.MaxFiles) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				lg.MaxFiles = n
-				return nil
-			},
-		},
-		{
-			Label: "Max Size KB", Help: "Rotation threshold in KB (Size type only)", Type: ftInteger, Col: 3, Row: 6, Width: 7, Min: 1, Max: 1048576,
-			Get: func() string { return strconv.Itoa(lg.MaxSizeKB) },
-			Set: func(val string) error {
-				n, err := strconv.Atoi(val)
-				if err != nil {
-					return err
-				}
-				lg.MaxSizeKB = n
-				return nil
-			},
-		},
 	}
 }
